+++ src/expr/lexer.rs
@@ -24,13 +24,14 @@ pub enum Token<'a> {
     Comma,
 }
 
-impl<'a> std::fmt::Display for Token<'a> {
+impl std::fmt::Display for Token<'_> {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         std::fmt::Debug::fmt(self, f)
     }
 }
 
-impl<'a> Token<'a> {
+impl Token<'_> {
+    #[inline]
     fn len(&self) -> usize {
         match self {
             Token::Key(s) => s.len(),